implementing recording storage

+++ spike/android-testing/src/uk/ac/aber/cs221/storage/SpeciesStorage.java
@@ -17,61 +17,59 @@ public class SpeciesStorage extends Storage<Species> {
    
    SpeciesStorage(Context context) {
       instance = this;
-       database = new DatabaseHelper(context);
+      database = new DatabaseHelper(context);
    }
    
    public static SpeciesStorage getInstance(Context context) {
-
-       if (instance == null) {
-           instance = new SpeciesStorage(context);
-       }
-       return instance;
+      
+      if (instance == null) {
+         instance = new SpeciesStorage(context);
+      }
+      return instance;
    }
-
    
    @Override
    public Cursor getCursor() {
       SQLiteDatabase connection = database.getReadableDatabase();
-       Cursor c = connection.query(table, null, null, null, null, null ,null);
-       database.close();
-       return c;
-
+      Cursor c = connection.query(table, null, null, null, null, null, null);
+      database.close();
+      return c;
+      
    }
    
    @Override
    public long store(Species species) {
-       ContentValues values = new ContentValues();
-       values.put("name", species.name);
-       values.put("comment", species.comment);
-       values.put("image_1", species.imageFile1);
-       values.put("image_2", species.imageFile2);
-       values.put("latitude", species.loc.getLatitude());
-       values.put("longitude", species.loc.getLongitude());
+      ContentValues values = new ContentValues();
+      values.put("name", species.name);
+      values.put("comment", species.comment);
+      values.put("image_1", species.imageFile1);
+      values.put("image_2", species.imageFile2);
+      values.put("latitude", species.loc.getLatitude());
+      values.put("longitude", species.loc.getLongitude());
       // name text not null, comment text not null, image_1 text not null,
       // image_2 text not null, latitude real, longitude real
       return database.getWritableDatabase().insert(table, null, values);
-
+      
    }
    
    @Override
-   public Species get(int id) {
+   public Species get(long id) {
       Cursor c = database.getReadableDatabase().rawQuery(
-              "SELECT * FROM " + table + " WHERE ID = " + id + " LIMIT 1;",
-              new String[] {});
-       if(c.getCount() > 0) {
-           Species species = new Species();
-
-           species.id = c.getInt(c.getColumnIndex("id"));
-
-       }
-       //toDO
-       return null;
-
+            "SELECT * FROM " + table + " WHERE ID = " + id + " LIMIT 1;",
+            new String[] {});
+      if (c.getCount() > 0) {
+         Species species = new Species();
+         
+         species.id = c.getInt(c.getColumnIndex("id"));
+         
+      }
+      // toDO
+      return null;
    }
    
    @Override
    public Species createNew() {
-
+      
       return null;
    }
 }
